SmartHome_AI_Camera/requirements.txt
opencv-python
ultralytics
face_recognition
pygame
pyyaml
requests
